break test_baseline into multiple files for easier parallel running
